+++ lib/src/point.dart
@@ -23,7 +23,7 @@ class Point {
   
   int get hashCode => row.hashCode ^ column.hashCode;
   
-  js.JsObject _toProxy() => _jsify({'row': row, 'column': column});
+  js.JsObject _toProxy() => _jsMap({'row': row, 'column': column});
   
   String toString() => 'Point: [${row}/${column}]';
 }