+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Optimized some list and map conversions on the `dart:js` boundary.
 - Added the `Annotation` class and expose the `onChangeAnnotation`, 
 `annotations`, and `clearAnnotations` members on the `EditSession` class.
+- Added a new `example/annotations`.
 
 ## 0.1.5+12.2.2013