+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Ace.dart Changes
 
-## 0.1.6-dev+12.2.2013
+## 0.1.6+12.2.2013
 
 - Added the `UndoManagerBase` class and implemented the 
 `EditSession.undoManager` setter; the argument to this setter _must_ be an